Safari-611.1.10/LayoutTests/fast/events/ios/dom-update-on-keydown-quirk-expected.txt

This tests that the value of the field is updated by the time any timer scheduled on keydown, keypress, or keyup fires. To run this test manually, focus the text field and press [.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



For keydown:
PASS document.getElementById("input").value is "["

For keypress:
PASS document.getElementById("input").value is "["

For input:
PASS document.getElementById("input").value is "["

For keyup:
PASS document.getElementById("input").value is "["
PASS successfullyParsed is true

TEST COMPLETE
